+++ resources/lang/en/custom-forms.php
@@ -1,0 +1,6 @@
+<?php
+
+// translations for Rupadana/FilamentCustomForms
+return [
+    //
+];