+++ src/FilamentCustomForms.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Rupadana\FilamentCustomForms;
+
+class FilamentCustomForms
+{
+}