+++ src/Testing/TestsFilamentCustomForms.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace VendorName\Skeleton\Testing;
+namespace Rupadana\FilamentCustomForms\Testing;
 
 use Livewire\Features\SupportTesting\Testable;
 
 /**
  * @mixin Testable
  */
-class TestsSkeleton
+class TestsFilamentCustomForms
 {
     //
 }