+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+All notable changes to `filament-custom-forms` will be documented in this file.
 
-## 1.0.0 - 202X-XX-XX
+## 3.0.0 - 2023-10-02
 
 - initial release
+- add InputGroup Component